lint fixes and renaming backend to backend.service
