+++ src/app/app.module.ts
@@ -35,7 +35,7 @@ import { FiltersComponent } from './components/filters/filters.component';
 import { CardComponent } from './components/card/card.component';
 import { FeaturesComponent } from './components/features/features.component';
 
-import { Backend } from './services/backend';
+import { Backend } from './services/backend.service';
 import { MaterialIconsService } from './services/material-icons/material-icons.service';
 import { Recipe } from './models/recipe.model';
 import { appReducer } from './store/reducers/reducers';
