+++ src/app/components/recipe-details/recipe-details.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input } from '@angular/core';
-import { Backend } from '../../services/backend';
+import { Backend } from '../../services/backend.service';
 import { ActivatedRoute } from '@angular/router';
 import 'rxjs/add/operator/mergeMap';
 import { State } from '../../models/state.model';
